refactor(db): extract conversation checkpoints into db/checkpoint.rs
